+++ src/lib/components/ui/link/link.svelte
@@ -8,7 +8,7 @@
 
 <a
 	{href}
-	class={cn('inline font-medium underline underline-offset-4', className)}
+	class={cn('font-medium underline underline-offset-4', className)}
 >
 	<slot />
 </a>